Dosctrings should follow a consistent predictable format.
Pick one you like and stick with it.  We use the reST standard, I'll leave a link in the slack chat for you to check out with examples of common standards.  If you are using python3.8 or higher, you should definitely be using type hints as well.  Type hints and standardized docstrings go a LONG way to helping `explain the code.`